Rewrote ISRs in assembly allowing -O3 compiling
Previously, the ISRs for the timer (and other interrupt
request handlers) were implemented in C. This compiled
and ran fine under GCC -O0 optimisation. However, with
optimisation, GCC no longer preserved the stack or
registers over the interrupt routine, and used ret
instead of iret. This lead to highly unpredictable
results, including invalid operand, general
protection fault and page fault errors upon
experimentation.

ISRs for hardware interrupts are now implemented in
assembly, preserving the stack and registers. Due to
the time taken to solve this issue, this commit
is simply saving the now working state.

TODO:
 - clean up files modified and created
 - implement all interrupts in isr.asm

Many files were modified in fixing this problem. Files
with logic changes are:
 - gdt and idt -- it was discover the inline assembly
   syntax used was incorrect. Though not the ultimate
   souce of the problem, proper 'extended asm' syntax
   is now used.
 - interrupts.c -- previous hardware interrupt handlers
   no longer used. IDG_HANDLER cast redundant and
   removed.
 - isr.asm and isr.h -- locations of new ISRs
 - keyboard driver -- parameter type changed to
   uint32_t as address size anyway.

Note gcc also supports a __attribute__((interrupt)),
though this was not used.

+++ os/drivers/keyboard.c
@@ -34,7 +34,7 @@ static char *scan_codes2[0x40] = {"?\0", "?\0", "1\0", "2\0", "3\0", "4\0", "5\0
 
 // covers 0x0 - 0x39 of original XT scan codes.
 // Called by keyboard IRQ1 handler in interrupts.c in os/cpu/
-void printKeyInput(uint8_t scan_code)
+void printKeyInput(uint32_t scan_code)
 {
     if (scan_code <= 0x39) {
         print(scan_codes2[scan_code]);
